requirements.txt
boto3
aws-xray-sdk
flask-lambda-python36
Flask==0.12.2
